tests: update unit and integration tests for SitesService
