+++ src/integration/Sites.spec.ts
@@ -10,7 +10,8 @@ import { SitesRouter as _SitesRouter } from "@root/routes/v2/authenticated/sites
 import { GitHubService } from "@root/services/db/GitHubService"
 import { ConfigYmlService } from "@root/services/fileServices/YmlFileServices/ConfigYmlService"
 import IsomerAdminsService from "@root/services/identity/IsomerAdminsService"
-import { SitesService } from "@root/services/utilServices/SitesService"
+import SitesService from "@root/services/identity/SitesService"
+import TokenStore from "@root/services/identity/TokenStore"
 import { getUsersService } from "@services/identity"
 import { sequelize } from "@tests/database"
 
@@ -25,11 +26,14 @@ const gitHubService = new GitHubService({ axiosInstance: mockAxios.create() })
 const configYmlService = new ConfigYmlService({ gitHubService })
 const usersService = getUsersService(sequelize)
 const isomerAdminsService = new IsomerAdminsService({ repository: IsomerAdmin })
+const tokenStore = new TokenStore()
 const sitesService = new SitesService({
+  siteRepository: Site,
   gitHubService,
   configYmlService,
   usersService,
   isomerAdminsService,
+  tokenStore,
 })
 
 const SitesRouter = new _SitesRouter({ sitesService })
